requirements.txt
fastapi
uvicorn[standard]
llama-cpp-python==0.2.36
